Logs: normalize field names
