+++ public/app/features/logs/components/panel/LogLineDetailsDisplayedFields.tsx
@@ -6,12 +6,10 @@ import { GrafanaTheme2 } from '@grafana/data';
 import { t } from '@grafana/i18n';
 import { Card, IconButton, useStyles2 } from '@grafana/ui';
 
-import { LOG_LINE_BODY_FIELD_NAME } from '../LogDetailsBody';
-import { LOG_LINE_ATTRIBUTES_FIELD_NAME } from '../otel/formats';
-
 import { LogLineDetailsMode } from './LogLineDetails';
 import { useLogListContext } from './LogListContext';
 import { reportInteractionOnce } from './analytics';
+import { getNormalizedFieldName } from './processing';
 
 export const LogLineDetailsDisplayedFields = () => {
   const { displayedFields, setDisplayedFields } = useLogListContext();
@@ -99,9 +97,7 @@ const DisplayedField = ({
     <div ref={provided.innerRef} {...provided.draggableProps} {...provided.dragHandleProps}>
       <Card noMargin className={styles.fieldCard}>
         <div className={styles.fieldWrapper}>
-          <div className={styles.field}>
-            {getNormalizedFieldName(field)}
-          </div>
+          <div className={styles.field}>{getNormalizedFieldName(field)}</div>
           {displayedFields.length > 1 && (
             <>
               <IconButton
@@ -149,12 +145,3 @@ const getStyles = (theme: GrafanaTheme2, detailsMode: LogLineDetailsMode) => ({
     flex: 1,
   }),
 });
-
-export function getNormalizedFieldName(field: string) {
-  if (field === LOG_LINE_BODY_FIELD_NAME) {
-    return t('logs.log-line-details.log-line-field', 'Log line');
-  } else if (field === LOG_LINE_ATTRIBUTES_FIELD_NAME) {
-    return t('logs.log-line-details.log-attributes-field', 'Log attributes');
-  }
-  return field;
-}
